% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asumsi.R
\name{autokorelasi}
\alias{autokorelasi}
\title{Uji Durbin-Watson Manual}
\usage{
autokorelasi(model)
}
\arguments{
\item{model}{Hasil dari `ols_fit()`.}
}
\value{
Statistik DW dan interpretasi sederhana.
}
\description{
Menghitung statistik DW untuk mendeteksi autokorelasi.
}
