% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asumsi.R
\name{homoskedasticity}
\alias{homoskedasticity}
\title{Uji Homoskedastisitas Secara Visual}
\usage{
homoskedasticity(model)
}
\arguments{
\item{model}{Hasil dari `ols_fit()`.}
}
\description{
Pemeriksaan pola residual secara visual dan simpulan sederhana.
}
