% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asumsi.R
\name{multikolinearitas}
\alias{multikolinearitas}
\title{VIF Manual}
\usage{
multikolinearitas(X)
}
\arguments{
\item{X}{Matriks desain (sudah tanpa kolom intercept).}
}
\value{
Nilai VIF tiap kolom X.
}
\description{
Menghitung VIF tiap variabel prediktor secara manual.
}
