% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asumsi.R
\name{normality}
\alias{normality}
\title{Uji Normalitas Residual Secara Manual}
\usage{
normality(model)
}
\arguments{
\item{model}{Hasil dari `ols_fit()`.}
}
\value{
Pesan kesimpulan normalitas.
}
\description{
Menggunakan skewness dan kurtosis (metode moment).
}
