% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_fit.R
\name{ols_fit}
\alias{ols_fit}
\title{Estimasi Regresi Linier Berganda}
\usage{
ols_fit(X, y)
}
\arguments{
\item{X}{Matriks desain, kolom pertama biasanya adalah 1 (intercept).}

\item{y}{Vektor respons.}
}
\value{
List: koefisien, fitted values, residual, dan sigma2.
}
\description{
Menghitung koefisien regresi menggunakan metode OLS.
}
