% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ols.R
\name{predict_ols}
\alias{predict_ols}
\title{Prediksi dengan Model OLS}
\usage{
predict_ols(model, X_new)
}
\arguments{
\item{model}{List dari hasil `ols_fit()`.}

\item{X_new}{Matriks desain baru.}
}
\value{
Vektor nilai prediksi.
}
\description{
Prediksi nilai baru berdasarkan model OLS.
}
